删除 router,去除umi依赖
